add another decision page in case of a disagreement
When there is a disagreement amongst the team we add an option to chat. If you reach an agreement at this time you get 1 cent bonus. If your original decision is picked by the team, you get 3 cents; the same amount as you would get when everyone picked the same sense initially